+++ app/src/main/java/com/rikvanvelzen/codingtest/data/models/dto/CurrencyRatesDTO.kt
@@ -1,10 +1,12 @@
 package com.rikvanvelzen.codingtest.data.models.dto
 
+import java.math.BigDecimal
+
 data class CurrencyRatesDTO(
 
         val base: String? = null,
         val date: String? = null,
-        val rates: Map<String, Double>? = null
+        val rates: Map<String, BigDecimal>? = null
 )
